passes: Move implementations into implementation files
It is not longer necessary to have the full Visitor class specifications
inside the headers. This reduces the dependency surface, and makes the
abstractions cleaner. This is part of the plan outlined in #3827.


stack-info: PR: https://github.com/bpftrace/bpftrace/pull/3888, branch: user/amscanne/diags/14
