+++ src/ast/passes/config_analyser.cpp
@@ -1,13 +1,52 @@
-#include "config_analyser.h"
-
 #include <cstring>
 #include <string>
 
 #include "ast/ast.h"
+#include "ast/passes/config_analyser.h"
+#include "ast/visitor.h"
+#include "bpftrace.h"
 #include "config.h"
 #include "types.h"
 
 namespace bpftrace::ast {
+
+namespace {
+
+class ConfigAnalyser : public Visitor<ConfigAnalyser> {
+public:
+  explicit ConfigAnalyser(BPFtrace &bpftrace)
+      : bpftrace_(bpftrace),
+        config_setter_(ConfigSetter(*bpftrace.config_, ConfigSource::script))
+  {
+  }
+
+  using Visitor<ConfigAnalyser>::visit;
+  void visit(Integer &integer);
+  void visit(String &string);
+  void visit(StackMode &mode);
+  void visit(AssignConfigVarStatement &assignment);
+
+private:
+  BPFtrace &bpftrace_;
+  ConfigSetter config_setter_;
+
+  void set_config(AssignConfigVarStatement &assignment, ConfigKeyInt key);
+  void set_config(AssignConfigVarStatement &assignment, ConfigKeyBool key);
+  void set_config(AssignConfigVarStatement &assignment, ConfigKeyString key);
+  void set_config(AssignConfigVarStatement &assignment,
+                  ConfigKeyUserSymbolCacheType key);
+  void set_config(AssignConfigVarStatement &assignment,
+                  ConfigKeySymbolSource key);
+  void set_config(AssignConfigVarStatement &assignment, ConfigKeyStackMode key);
+  void set_config(AssignConfigVarStatement &assignment,
+                  ConfigKeyMissingProbes key);
+
+  void log_type_error(SizedType &type,
+                      Type expected_type,
+                      AssignConfigVarStatement &assignment);
+};
+
+} // namespace
 
 void ConfigAnalyser::log_type_error(SizedType &type,
                                     Type expected_type,
