+++ src/ast/passes/pid_filter_pass.cpp
@@ -1,8 +1,26 @@
-#include "pid_filter_pass.h"
-
+#include "ast/passes/pid_filter_pass.h"
 #include "ast/ast.h"
+#include "ast/visitor.h"
+#include "bpftrace.h"
 
 namespace bpftrace::ast {
+
+namespace {
+
+class PidFilterPass : public Visitor<PidFilterPass> {
+public:
+  explicit PidFilterPass(ASTContext &ast, BPFtrace &bpftrace)
+      : ast_(ast), bpftrace_(bpftrace)
+  {
+  }
+
+  using Visitor<PidFilterPass>::visit;
+  void visit(Probe &probe);
+
+private:
+  ASTContext &ast_;
+  BPFtrace &bpftrace_;
+};
 
 // If the probe can't filter by pid when attaching
 // then we inject custom AST to filter by pid.
@@ -38,6 +56,8 @@ bool probe_needs_pid_filter(AttachPoint *ap)
 
   return false;
 }
+
+} // namespace
 
 static Statement *create_pid_filter(ASTContext &ast,
                                     int pid,
