+++ src/ast/passes/pid_filter_pass.h
@@ -1,25 +1,8 @@
 #pragma once
 
 #include "ast/pass_manager.h"
-#include "ast/visitor.h"
-#include "bpftrace.h"
 
 namespace bpftrace::ast {
-
-class PidFilterPass : public Visitor<PidFilterPass> {
-public:
-  explicit PidFilterPass(ASTContext &ast, BPFtrace &bpftrace)
-      : ast_(ast), bpftrace_(bpftrace)
-  {
-  }
-
-  using Visitor<PidFilterPass>::visit;
-  void visit(Probe &probe);
-
-private:
-  ASTContext &ast_;
-  BPFtrace &bpftrace_;
-};
 
 Pass CreatePidFilterPass();
