+++ src/ast/passes/probe_analyser.cpp
@@ -1,6 +1,27 @@
 #include "probe_analyser.h"
+#include "ast/visitor.h"
+#include "bpftrace.h"
 
 namespace bpftrace::ast {
+
+namespace {
+
+class ProbeAnalyser : public Visitor<ProbeAnalyser> {
+public:
+  explicit ProbeAnalyser(ASTContext &ctx, BPFtrace &bpftrace)
+      : ctx_(ctx), bpftrace_(bpftrace)
+  {
+  }
+
+  using Visitor<ProbeAnalyser>::visit;
+  void visit(Probe &probe);
+
+  void analyse();
+
+private:
+  const ASTContext &ctx_;
+  const BPFtrace &bpftrace_;
+};
 
 void ProbeAnalyser::visit(Probe &probe)
 {
@@ -37,6 +58,8 @@ void ProbeAnalyser::visit(Probe &probe)
     }
   }
 }
+
+} // namespace
 
 Pass CreateProbePass()
 {
